# backend/cmd/main/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Petplace API Description",
        "title": "Petplace API Version1",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/api",
    "paths": {
...
        "/cageroom": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "create cage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CageRooms"
                ],
                "summary": "Create Cage",
                "parameters": [
                    {
                        "description": "cageroom payload",
                        "name": "CageRoom",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CageRoom"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/cageroom/all/{profile_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get cages",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CageRooms"
                ],
                "summary": "Get Cages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Profile ID",
                        "name": "profile_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/cageroom/search": {
            "get": {
                "description": "get selected cages",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CageRooms"
                ],
                "summary": "Get Selected Cages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by animal type and size ex. animals[0].animal_type=dog\u0026animals[0].cage_size=s",
                        "name": "animals",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by start_time",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by end_time",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by latitude",
                        "name": "latitude",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by longitude",
                        "name": "longitude",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by sort",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/cageroom/search/{user_id}/{profile_id}": {
            "get": {
                "description": "get selected cages by hotel",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CageRooms"
                ],
                "summary": "Get Selected Cages by Hotel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID of Client",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Profile ID",
                        "name": "profile_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by animal type and size ex. animals[0].animal_type=dog\u0026animals[0].cage_size=s",
                        "name": "animals",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by start_time",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by end_time",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by latitude",
                        "name": "latitude",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by longitude",
                        "name": "longitude",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by sort",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/cageroom/type/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Cage type and animal type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CageRooms"
                ],
                "summary": "Get Cage type and animal type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/cageroom/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get cage",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CageRooms"
                ],
                "summary": "Get Cage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "update cage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CageRooms"
                ],
                "summary": "Update Cage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "cageroom payload",
                        "name": "CageRoom",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CageRoom"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "delete cage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CageRooms"
                ],
                "summary": "Delete Cage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CageRoom": {
            "type": "object",
            "required": [
                "cage_type",
                "height",
                "lenth",
                "max_capacity",
                "price",
                "profile_id",
                "quantity",
                "width"
            ],
            "properties": {
                "animal_type": {
                    "type": "string"
                },
                "cage_type": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "facility": {
                    "type": "string"
                },
                "facility_array": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "height": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "image_array": {
                    "description": "no in database",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "lenth": {
                    "description": "size",
                    "type": "number"
                },
                "max_capacity": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "profile_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "size": {
                    "type": "string"
                },
                "width": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}